Refactor: Create jetpack package and move classes
